# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Top1000 站点数据查询 API",
        "title": "Top1000 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:7066",
    "basePath": "/",
    "paths": {
        "/sites.json": {
            "get": {
                "description": "获取IYUU站点列表数据（需要配置IYUU_SIGN环境变量）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sites"
                ],
                "summary": "获取IYUU站点列表",
                "responses": {
                    "200": {
                        "description": "站点列表数据",
                        "schema": {}
                    },
                    "500": {
                        "description": "error\": \"无法加载站点数据",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "502": {
                        "description": "error\": \"未配置IYUU_SIGN环境变量",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/top1000.json": {
            "get": {
                "description": "获取Top1000站点列表数据，数据会自动更新（24小时过期）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Top1000"
                ],
                "summary": "获取Top1000站点数据",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ProcessedData"
                        }
                    },
                    "500": {
                        "description": "error\": \"无法加载数据",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.ProcessedData": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SiteItem"
                    }
                },
                "time": {
                    "type": "string"
                }
            }
        },
        "model.SiteItem": {
            "type": "object",
            "properties": {
                "duplication": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "siteName": {
                    "type": "string"
                },
                "siteid": {
                    "type": "string"
                },
                "size": {
                    "type": "string"
                }
            }
        }
    }
}